requirements.txt
pyperclip==1.8.2
PyYAML==5.4.1
sqlparse==0.4.1
tabulate==0.8.9
pymysql==1.0.2
pyodbc==4.0.31